# web/openapi.json
{
  "components": {
    "schemas": {
...
      "GraphEdge": {
        "description": "A directed, labelled relation from one node to another.",
        "properties": {
          "approved": {
            "default": true,
            "title": "Approved",
            "type": "boolean"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "origin": {
            "default": "seed",
            "maxLength": 8,
            "title": "Origin",
            "type": "string"
          },
          "relation": {
            "default": "related",
            "maxLength": 40,
            "title": "Relation",
            "type": "string"
          },
          "source_id": {
            "title": "Source Id",
            "type": "integer"
          },
          "target_id": {
            "title": "Target Id",
            "type": "integer"
          },
          "weight": {
            "default": "1",
            "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,4}|(?=[\\d.]{1,7}0*$)\\d{0,4}\\.\\d{0,2}0*$)",
            "title": "Weight",
            "type": "string"
          }
        },
        "required": [
          "source_id",
          "target_id"
        ],
        "title": "GraphEdge",
        "type": "object"
      },
      "GraphNode": {
        "description": "A node: an asset, a product, or a concept.",
        "properties": {
          "icon": {
            "anyOf": [
              {
                "maxLength": 32,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon"
          },
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "kind": {
            "default": "concept",
            "maxLength": 16,
            "title": "Kind",
            "type": "string"
          },
          "label": {
            "maxLength": 120,
            "title": "Label",
            "type": "string"
          },
          "symbol": {
            "anyOf": [
              {
                "maxLength": 24,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Symbol"
          }
        },
        "required": [
          "label"
        ],
        "title": "GraphNode",
        "type": "object"
      },
      "GraphView": {
        "description": "The connections graph \u2014 nodes and the edges between them.",
        "properties": {
          "edges": {
            "items": {
              "$ref": "#/components/schemas/GraphEdge"
            },
            "title": "Edges",
            "type": "array"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/GraphNode"
            },
            "title": "Nodes",
            "type": "array"
          }
        },
        "required": [
          "nodes",
          "edges"
        ],
        "title": "GraphView",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NodeCreate": {
        "properties": {
          "icon": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon"
          },
          "kind": {
            "default": "concept",
            "title": "Kind",
            "type": "string"
          },
          "label": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Label",
            "type": "string"
          },
          "symbol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Symbol"
          }
        },
        "required": [
          "label"
        ],
        "title": "NodeCreate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "summary": "Self-hosted automated Binance trading engine.",
    "title": "Capital Engine",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/connections/edges/{edge_id}": {
      "delete": {
        "description": "Delete an edge (reject a suggestion or prune a curated link).",
        "operationId": "delete_edge_api_connections_edges__edge_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "edge_id",
            "required": true,
            "schema": {
              "title": "Edge Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Delete Edge",
        "tags": [
          "connections"
        ]
      }
    },
    "/api/connections/edges/{edge_id}/approve": {
      "post": {
        "description": "Approve a pending edge so it joins the graph.",
        "operationId": "approve_edge_api_connections_edges__edge_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "edge_id",
            "required": true,
            "schema": {
              "title": "Edge Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphEdge"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Approve Edge",
        "tags": [
          "connections"
        ]
      }
    },
    "/api/connections/graph": {
      "get": {
        "description": "The full graph. Pending (AI-suggested) edges are opt-in.",
        "operationId": "get_graph_api_connections_graph_get",
        "parameters": [
          {
            "in": "query",
            "name": "include_pending",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Pending",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Graph",
        "tags": [
          "connections"
        ]
      }
    },
    "/api/connections/nodes": {
      "post": {
        "description": "Create a node (or return the existing one with the same label).",
        "operationId": "create_node_api_connections_nodes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NodeCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphNode"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create Node",
        "tags": [
          "connections"
        ]
      }
    },
    "/api/connections/suggest/{node_id}": {
      "post": {
        "description": "Ask the LLM to suggest related entities for a node (stored pending).",
        "operationId": "suggest_connections_api_connections_suggest__node_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "node_id",
            "required": true,
            "schema": {
              "title": "Node Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GraphEdge"
                  },
                  "title": "Response Suggest Connections Api Connections Suggest  Node Id  Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Suggest Connections",
        "tags": [
          "connections"
        ]
      }
    },
...
  }
}